📦 refactor(types): move types to src/types.ts
